Structure the README.md with some info
If you want to colloborate with someone on your code you should
define and agree on some basics.  The README is updated to show
what that might look like.

This commit message is structured as another form of comms that shows
me explaining why I have implemented this change rather than _what_
has changed.  Reading the code will tell you what changed, thus
this message should be the "meta" on why it was changed.